requirements.django.txt
Django==1.10
argparse==1.2.1
django-filter==0.14.0
djangorestframework==3.4.0
jsonfield==1.0.3
wsgiref==0.1.2
Pillow>=3.4.2
git+https://github.com/gerasev-kirill/django-gcloud-storage.git
